feat(comment-routes): create new files for comment routes
created new files to handle comments
