+++ api/server.js
@@ -3,6 +3,7 @@ const cors = require('cors');
 
 const authUser = require("../routes/authUser");
 const userPost = require("../routes/postRoute");
+const postComment = require("../routes/commentsRoute");
 
 const server = express();
 
@@ -11,5 +12,6 @@ server.use(cors());
 
 server.use('/api', authUser);
 server.use('/post', userPost);
+server.use("/comment", postComment);
 
 module.exports = server;
