+++ models/postModal.js
@@ -21,15 +21,6 @@ async function updatePost(user_id, post_id, changes) {
    return db("posts as p").where("p.id", post.post_id).update(changes,"id");
 }
 
-// COMMENTS
-
-async function addComment(user_id, post_id, comment) {
-   const [id] = await db('comments').insert(comment, 'id');
-   const userCommentBody = {user_id, post_id, comment_id: id};
-   const commentId = await db("user_comment").insert(userCommentBody, 'id');
-   return commentId;
-}
-
 async function getAll() {
    const posts = await db("posts");
    return Promise.all(
@@ -40,26 +31,9 @@ async function getAll() {
    )
 }
 
-function findUserComments(user_id, comment_id) {
-   return  db("comments as c").join("user_comment as uc", "c.id", "uc.comment_id").where({"uc.user_id": user_id}).where({"uc.comment_id": comment_id});
-}
-
-async function removeComment(user_id, comment_id) {
-   const [comment] = await findUserComments(user_id, comment_id);
-   return db("comments as c").where("c.id", comment.id).del();
-}
-
-async function updateComment(user_id, comment_id, changes) {
-   const [comment] = await findUserComments(user_id, comment_id);
-   return db("comments as c").where("c.id", comment.id).update(changes);
-}
-
 module.exports = {
    add,
    getAll,
    removePost,
-   updatePost,
-   addComment,
-   removeComment,
-   updateComment
+   updatePost
 };
